+++ OrmasBL/OrmasBL/OrderListClass.h
@@ -10,12 +10,12 @@ namespace BusinessLayer
 		int id = 0;
 		int orderID = 0;
 		int productID = 0;
-		int count = 0;
+		double count = 0;
 		double sum = 0;
 		int statusID = 0;
 		int currencyID = 0;
 	public:
-		OrderList(int olID, int oID,  int pID, int olCount, double olSum, int sID, int cID) :id(olID), orderID(oID), productID(pID), 
+		OrderList(int olID, int oID, int pID, double olCount, double olSum, int sID, int cID) :id(olID), orderID(oID), productID(pID),
 			count(olCount), sum(olSum), statusID(sID), currencyID(cID){};
 		OrderList(DataLayer::orderListCollection);
 		OrderList(){};
@@ -25,7 +25,7 @@ namespace BusinessLayer
 		int GetID();
 		int GetOrderID();
 		int GetProductID();
-		int GetCount();
+		double GetCount();
 		double GetSum();
 		int GetStatusID();
 		int GetCurrencyID();
@@ -34,7 +34,7 @@ namespace BusinessLayer
 		void SetID(int);
 		void SetOrderID(int);
 		void SetProductID(int);
-		void SetCount(int);
+		void SetCount(double);
 		void SetSum(double);
 		void SetStatusID(int);
 		void SetCurrencyID(int);
@@ -44,9 +44,9 @@ namespace BusinessLayer
 		bool UpdateOrderList(DataLayer::OrmasDal& ormasDal, std::string& errorMessage);
 		bool DeleteOrderList(DataLayer::OrmasDal& ormasDal, std::string& errorMessage);
 		bool DeleteListByOrderID(DataLayer::OrmasDal& ormasDal, int oID, std::string& errorMessage);
-		bool CreateOrderList(DataLayer::OrmasDal& ormasDal, int oID, int pID, int olCount, double olSum,
+		bool CreateOrderList(DataLayer::OrmasDal& ormasDal, int oID, int pID, double olCount, double olSum,
 			int sID, int cID, std::string& errorMessage);
-		bool UpdateOrderList(DataLayer::OrmasDal& ormasDal, int oID,int pID, int olCount, double olSum,
+		bool UpdateOrderList(DataLayer::OrmasDal& ormasDal, int oID, int pID, double olCount, double olSum,
 			int sID, int cID, std::string& errorMessage);
 
 		//Generate filter string for class
@@ -55,7 +55,7 @@ namespace BusinessLayer
 		bool IsEmpty();
 		void Clear();
 	private:
-		bool IsDuplicate(DataLayer::OrmasDal& ormasDal, int oID, int pID, int olCount, double olSum,
+		bool IsDuplicate(DataLayer::OrmasDal& ormasDal, int oID, int pID, double olCount, double olSum,
 			 int cID, std::string& errorMessage);
 		bool IsDuplicate(DataLayer::OrmasDal& ormasDal, std::string& errorMessage);
 	};
